TipTapEditor 리팩토링
